requirements.txt
# GUI
pyside6
PyOpenGL == 3.1.9a2
PyOpenGL_accelerate == 3.1.9a2
pywavefront
# 逻辑
numpy
pillow
PyYAML
# minecraft
minecraft-launcher-lib
